@@ -8,29 +8,32 @@
     "build": "ng build --prod",
     "test": "ng test",
     "lint": "ng lint",
-    "e2e": "ng e2e"
+    "e2e": "ng e2e",
+    "build:server:prod": "ng run my-shop:server && webpack --config webpack.server.config.js --progress --colors",
+    "build:browser:prod": "ng build --prod",
+    "build:prod": "npm run build:server:prod && npm run build:browser:prod",
+    "server": "node local.js"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^5.2.11",
+    "@angular/animations": "^6.1.10",
     "@angular/cdk": "^5.2.5",
-    "@angular/common": "^5.2.0",
-    "@angular/compiler": "^5.2.0",
-    "@angular/core": "^5.2.0",
-    "@angular/flex-layout": "^5.0.0-beta.14",
-    "@angular/forms": "^5.2.0",
-    "@angular/http": "^5.2.0",
+    "@angular/common": "^6.1.10",
+    "@angular/compiler": "^6.1.10",
+    "@angular/core": "^6.1.10",
+    "@angular/flex-layout": "^6.0.0-beta.18",
+    "@angular/forms": "^6.1.10",
+    "@angular/http": "^6.1.10",
     "@angular/material": "^5.2.5",
-    "@angular/platform-browser": "^5.2.0",
-    "@angular/platform-browser-dynamic": "^5.2.0",
-    "@angular/router": "^5.2.0",
+    "@angular/platform-browser": "^6.0.0",
+    "@angular/platform-browser-dynamic": "^6.1.10",
+    "@angular/router": "^6.1.10",
     "@ng-bootstrap/ng-bootstrap": "^2.0.0",
-    "@ng-toolkit/universal": "^8.0.3",
+    "@ng-toolkit/universal": "^1.1.40",
     "@swimlane/ngx-datatable": "^14.0.0",
     "admin-lte": "^2.4.8",
     "angular-admin-lte": "^1.2.7",
     "angular-jwt": "^0.1.10",
-    "angular2-jwt": "^0.2.3",
     "bootstrap": "^4.3.1",
     "core-js": "^2.4.1",
     "ez-plus": "^1.2.1",
@@ -43,16 +46,23 @@
     "ngx-slider": "0.0.5",
     "npm": "^6.13.0",
     "popper.js": "^1.14.3",
-    "rxjs": "^5.5.6",
+    "rxjs": "^6.5.3",
+    "rxjs-compat": "^6.5.3",
+    "rxjs-tslint": "^0.1.7",
     "semantic-ui": "^2.4.2",
     "start": "^5.1.0",
     "tether": "^1.4.4",
-    "zone.js": "^0.8.19"
+    "zone.js": "^0.8.29",
+    "@angular/platform-server": "^6.0.0",
+    "@nguniversal/module-map-ngfactory-loader": "^6.0.0",
+    "@nguniversal/express-engine": "^6.0.0",
+    "cors": "~2.8.4"
   },
   "devDependencies": {
-    "@angular/cli": "~1.7.3",
-    "@angular/compiler-cli": "^5.2.0",
-    "@angular/language-service": "^5.2.0",
+    "@angular-devkit/build-angular": "~0.8.0",
+    "@angular/cli": "^6.2.9",
+    "@angular/compiler-cli": "^6.1.10",
+    "@angular/language-service": "^6.1.10",
     "@types/jasmine": "~2.8.3",
     "@types/jasminewd2": "~2.0.2",
     "@types/jquery": "^3.3.1",
@@ -66,8 +76,10 @@
     "karma-jasmine": "~1.1.0",
     "karma-jasmine-html-reporter": "^0.2.2",
     "protractor": "~5.1.2",
+    "ts-loader": "^4.2.0",
     "ts-node": "~4.1.0",
     "tslint": "~5.9.1",
-    "typescript": "~2.5.3"
+    "typescript": "~2.9.2",
+    "webpack-cli": "^2.1.5"
   }
 }
