@@ -1,3 +1,4 @@
+import { AppBrowserModule } from '.././src/app/app.browser.module';
 import { enableProdMode } from '@angular/core';
 import { platformBrowserDynamic } from '@angular/platform-browser-dynamic';
 
@@ -8,5 +9,5 @@ if (environment.production) {
   enableProdMode();
 }
 
-platformBrowserDynamic().bootstrapModule(AppModule)
+platformBrowserDynamic().bootstrapModule(AppBrowserModule)
   .catch(err => console.log(err));
